default/requirements.txt
Flask==0.10.1
gcloud==0.14.0
gunicorn==19.6.0
oauth2client==2.1.0
Flask-SQLAlchemy==2.1
PyMySQL==0.7.3
Flask-PyMongo==0.4.1
PyMongo==3.2.2
six==1.10.0
requests[security]==2.10.0
honcho==0.7.1
psq==0.3.0
redis
google-cloud-logging
